+++ src/Service/ControllerManagerFactory.php
@@ -11,8 +11,7 @@ namespace Zend\Mvc\Service;
 
 use Interop\Container\ContainerInterface;
 use Zend\Mvc\Controller\ControllerManager;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
 
 class ControllerManagerFactory implements FactoryInterface
 {
@@ -37,18 +36,5 @@ class ControllerManagerFactory implements FactoryInterface
             return new ControllerManager($container, $options);
         }
         return new ControllerManager($container);
-    }
-
-    /**
-     * Create and return ControllerManager instance
-     *
-     * For use with zend-servicemanager v2; proxies to __invoke().
-     *
-     * @param ServiceLocatorInterface $container
-     * @return ControllerManager
-     */
-    public function createService(ServiceLocatorInterface $container)
-    {
-        return $this($container, ControllerManager::class);
     }
 }