+++ src/Service/ViewTemplateMapResolverFactory.php
@@ -10,8 +10,7 @@
 namespace Zend\Mvc\Service;
 
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
 use Zend\View\Resolver as ViewResolver;
 
 class ViewTemplateMapResolverFactory implements FactoryInterface
@@ -38,18 +37,5 @@ class ViewTemplateMapResolverFactory implements FactoryInterface
             }
         }
         return new ViewResolver\TemplateMapResolver($map);
-    }
-
-    /**
-     * Create and return ViewResolver\TemplateMapResolver instance
-     *
-     * For use with zend-servicemanager v2; proxies to __invoke().
-     *
-     * @param ServiceLocatorInterface $container
-     * @return ViewResolver\TemplateMapResolver
-     */
-    public function createService(ServiceLocatorInterface $container)
-    {
-        return $this($container, ViewResolver\TemplateMapResolver::class);
     }
 }